lms_complete_final_mariadb_updated_2 (2)/lms_backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Pillow==10.0.0
redis==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

PyMySQL==1.1.0

